tp3/requirements.txt
requests
pymongo
python-dotenv
psycopg2-binary
